ui: move tags components to root of component directories
This commit moves the tags components directory, because the idea of ​​the
project to keep the components in the root.
