+++ ui/src/components/device/DeviceList.vue
@@ -140,7 +140,7 @@
 import TerminalDialog from '@/components/terminal/TerminalDialog';
 import DeviceIcon from '@/components/device/DeviceIcon';
 import DeviceDelete from '@/components/device/DeviceDelete';
-import TagFormUpdate from '@/components/setting/tag/TagFormUpdate';
+import TagFormUpdate from '@/components/tag/TagFormUpdate';
 import { lastSeen } from '@/components/filter/date';
 import formatDeviceSort from '@/components/filter/object';
